Remove unused no-installed-files test feature
Looking at 45c8557d, the idea behind this seems to be that a test could
conditionally indicate that the list of installed files should not be
validated by creating that file.

It's no longer used anywhere.

Also remove a lingering no-install-files file which isn't used since commit
c693bd9b.