Redetect machines after every change to languages
Currently, detect_machine_info() is called:

- from Environment.__init__(), before we have any compiler available

- from Interpreter.__init__() with the list of languages provided to
project() (determined via the initial parse_project())

This is not sufficent in the case where no languages are specified to
project() and are later added with add_languages().

We cannot correctly detect that the host machine should be treated as
x86 (on x86_64 hardware) until we have a compiler we are told to use.

Redetect machines after project(), and after every add_languages().