Test host_machine is correctly detected after add_languages()
Test that the host_machine is correctly detected after add_languages(),
when no langauge is initially specified in project().

In the MSYS2 MSYSTEM=MINGW32 environment (64-bit MSYS2 but with a
i686-w64-mingw32 targeted gcc as gcc) this test fails, as it
(incorrectly) tries to build retval-x86_64.S using an x86 compiler.