modules: Add a wayland module
Currently this only adds a wrapper around wayland-scanner for handling
the setup and some of the trickiness of that tool. I noticed that a
number of projects (mesa, weston, gnome, a couple of test tools we have
internally) are all porting around different code for trying to find the
scanner and point at files in the protocol directory, and decided that a
module to handle this would be a welcome addition.