depdenencies/qt: Second attempt to allow qt --no-framework on MacOS.
Sometimes qt can be installed not as framework on MacOS. One way to
achieve this behaviour is to use conan package manager.

Allow falling back to simple library search if framework was
not found. In addition, allow to find the debug version of qt debug
libraries which have "_debug" suffix added to them.

Fixes #5091