run_project_tests: Fix it to actually work
This move the sorting logic into the TestDef class itself, which
simplifies sorting them.

Additionally it remove overcomplicated sort logic, because python
strings are compared character by character, we don't need to do the
split and cast to int, we know that realistically a maximum of 4
characters (the first 4 numbers) are going to be compared in most cases.