compilers: Make compile() and cached_compile() async
They now only start the compilation process but instead of blocking
until the result is known, they return a CompileContext object. Caller
is responsible for calling wait() on it to get the result.

This will allow to start multiple compilations in parallel and only
block on the result when every processes have been spawned.