boost: Rewrite the boost system dependency
This PR significantly improves the handling of the boost library
tags and also ensures that the found libraries are always compatible
(have the same ABI tag). The current setup can also be extended to
filter for additional features (static linking with the runtime, etc.).

Additionally, BOOST_ROOT is better supported (it is now guaranteed
that all found files belong to a single root directory).

Finally, boost.py is now fully annotated (and checked with mypy).