Exclude .hg* from dist tarballs
Debian's lintian checker complaints when upstream tarballs contain an
.hgtags file.

This excludes ".hg[a-z]*".  This is mostly consistent with the git
handling in del_gitfiles() which deletes .git*.  hg archive --help shows
an example of -X ".hg*".  However, instead of ".hg*", I have used
".hg[a-z]*" to keep the automatically added hg_archival.txt.  This file
may be useful to link the tarball to the Mercurial revision for either
manual inspection or in case any code interprets it for a --version or
similar.

This also excludes .hgignore and other things like .hgflow, which seems
desirable.

Fixes #6575