minstall: Fix check for when something was installed
If we preserve the existing file, we did not install anything. In that
case, the output now looks like:

```
ninja: Entering directory `.'
ninja: no work to do.
Nothing to install.
Preserved 1732 unchanged files, see meson-logs\install-log.txt for the full list
```

This also avoids an `isfile` lookup, which shaves off a few seconds in
gst-build install.