mesonbuild/compilers/mixins/clike.py: prefer .a library suffixes when --default-library=static
The existing library search order makes sense for environments with shared or a
mix of shared and static libraries. However, in a cross-compile environment
with a prebuilt cross-toolchain, both static and shared versions of the sysroot
libraries.

This environment presents a problem when a user builds an entirely static build
where there won't be any shared libraries at runtime because ld throws
the error: "ld: attempted static link of a dynamic object."