compilers: Don't put split soname args in start groups
Some compilers that act as linker drivers (dmd and ldc) need to split
arguments that GCC combines with , (ie, -Wl,-foo,bar -> -L=-foo -L=bar).
As such we need to detect that the previous argument contained -soname,
and not wrap that in a --start-group/--end-group

This modifies the shared library test to demonstrate the problem, with a
test case.

Fixes #6359