coredata: Set buildtype=release from others in more cases
`buildtype=release` has the semantic meaning 'release build' (f.ex.,
`b_ndebug=if-release`), and build files are using that to enable/disable
features.

However, that's broken when you want to set `-Doptimization` to
anything other than `3` for a 'release build', for example when
shipping for embedded or mobile, because build files will get
`custom`.

Setting `buildtype` to `release` when `optimization` is `2` or `s`
(and `debug=false`) is more useful than setting it to `custom`.