Add builtin docdir option
Projects that install documentation typically do this under
/usr/share/doc/$project, or any other prefix. This behaviour is
equivalent to the infodir option, which is already implemented.
It is the responsibility of the user, to specify a folder for his
(sub)project, as the default value of docdir will only point to the root
directory for documentation, /usr/share/doc, since there is no reliable
way for us to figure out the correct directory on a per project basis.