Site-wide native file configuration
This allows a special native file called `siteconfig.ini` to be placed
in a user or system wide location on Unix systems for default
configuration. My initial thought of it's usefulness is to allow distros
that use non standard lib locations to set those system wide in
configuration rather than us having to fix up libdir paths.

Ideally this would be followed up by a similar cross configuration
setting, but I'm unsure of how to do that, given the complexities of
cross compilation.