mtest: Avoid over commiting futures
This patch will prevent creating as many repeats as the total requested
number. This reduce memory pressure when a large number if requested and
will ensure that we stop processing if one of the parallel running test
failed.

Before this patch, we often endup queue all the futures before the first
failures and wait for all the futures in drain_futures(). This is not
very convenient when trying to reproduce a rarely failing test.