mparser: fix precedence of arithmetic operators
The arithmetic operators are now split into two groups:
    * The add/sub group: +, -
    * The mul/div group: *, /, %

All operators within the same group are left-associative and have equal
precedence. The mul/div group has a higher precedence than the add/sub
group, as one would expect.

Previously every operator had a different precedence and was
right-associative, which resulted in surprising behavior.

This is a potentially breaking change for projects that relied on the
old incorrect behavior.

Fixes #6870