dependencies/cmake: Suffix bare library dependencies on Windows
On Windows, library dependencies can be passed with no prefix or suffix;
rather than -lfoo or foo.dll, they can just be passed as 'foo'.

CMake handles this and suffixes the library with '.lib' or '.dll',
depending on the link mode.

Do the same here, and if we've been passed an unqualified non-option
bare name on Windows, add the appropriate suffix and pass it through to
the linker. This fixes dependencies on system libraries.