backend/vs: Fix build when not all languages have pch
It is not unheard-of for a project to use pch for C++ but not for C
(because C usually builds fast enough anyway, so it's not worth the
developer overhead of maintaining the pch file).

This code was trying to optimize the vcxproj file size by detecting
"only one language", but it was only looking at the number of
pch-languages defined. This is incorrect when pch is not defined for all
languages in use.

Instead of tweaking the optimization further, remove it. This makes the
vs backend behave more like the ninja backend.