dependency: Add dep.get_include_paths() function.
When dependencies are installed in odd or nonstandard locations,
external scripts such as those used in `custom_target` or `run_command`
can break.

A concrete example is [pybind's mkdoc.py][1], which uses
[clang.cindex][2] to scan a codebase and generate a header file
containing python docstrings for a C++ project. This takes similar
arguments to the clang compiler, and can fail with a fatal error if it
encounters an `#include<>` it cannot resolve.

This is currently achievable by adding a separate python script, but
that script ends up reimplementing the entire meson dependency
infrastructure, since it requires searching cmake, pkg-config and the
subprojects directory to be reliable across environments.

This PR adds a `dep.get_include_paths()` to dependency objects, providing
a list of folders for such tools to search for headers. This list can be
added to the argv passed to functions such as `run_command` and
`custom_target`.

  [1]: https://github.com/pybind/pybind11/blob/master/tools/mkdoc.py
  [2]: https://github.com/llvm-mirror/clang/blob/master/bindings/python/clang/cindex.py