WIP: gobject-introspection cross with pkg-config
gobject-introspection is infamous for being difficult to get working
with cross compilation. But the situation has improved recently with
https://gitlab.gnome.org/GNOME/gobject-introspection/-/merge_requests/64,
thanks to @kanavin from the Yocto project. The upshot is now only
*some*, not all, of the executables need to be run on the host platform
with the exe wrapper.

On the meson side, there have been two attempts to fix things for cross:

- https://github.com/mesonbuild/meson/pull/2965from @kanavin and the
Yocto projejct, which because it predates
https://github.com/mesonbuild/meson/pull/4010 had to be somewhat hacky

- https://github.com/mesonbuild/meson/pull/7072 recently merged which
allows specifying some binaries with a cross file.

But, I think we can make a more seamless user interface that won't
require extra config, like for native builds. gobject-introspection
provides the binaries in its pkg-config file, and Meson now cleanly
supports separate `native: true` and `native: false` pkg-config paths
and lookup.

We should just need to add separate `native: true` and `native: false`
deps, and I have done that, but I am not sure exactly which should be
used when. I am coming at this as a distro maintainer not even
particularly involved with gnome things and so I'll need some advice on
what to do next.