Add 'prefer_prebuilt' keyword for gtest/gmock dependencies
gmock and gtest can be built from source or included as prebuilt
dependencies. This only works when the method is declared as 'system',
but the detection is inflexible in letting the user choose between the
prebuilt lib and the source when both are present -- the prebuilt is
always favored. This change allows that to be overridden, preferring
the source before the prebuilt library.

Why is this useful? Exported C++ ABIs are never stable. Therefore,
prebuilt libraries that might exist in a distrbution can be strictly
incompatible with a meson-configured build. As an example, a distro
might generally build/package with libstdc++, but a project may want to
use -stdlib=libc++, which would result in linker errors when it comes
time to include the prebuilt gtest/gmock dependencies.