Allow passing an array to library(vs_module_defs)
There are some cases where we want to use a def file with one compiler,
but not another. For example, you might want to use one with msvc,
clang-cl, and intel-cl, but not with mingw. Previously this required
duplicating the library call, one for the compilers with the def file,
and one without. This allows us to pass an array, possibly of zero
elements.