backend/vs: Fix C/C++ intellisense and "Go to File Action"
Intellisense and "Go to File Action" breaks when you try to edit/view a header
file that include a header of one of the target dependencies, but, the target
builds without any issue.

The problem occurs because the field in the : [[ Project Property Pages  -->
C/C++  -->  General  -->  Additional Include Directories ]]
was not correctly set with the exported include directories of the target
dependencies.