Add --enable/--disable syntax sugar for features
Since the default 'feature' type makes the special names 'enabled' and
'disabled' significant, we have reason to justify hard-coding
--enable-foo and --disable-foo as -Dfoo=enabled and -Dfoo=disabled
respectively.

At the risk of increasing API inconsistency, I think this change is a
net gain because it:

- avoids breaking muscle memory or 'expectation' for new users coming to
  meson from autotools etc., thus increasing ease of transitioning.

- removes one of the barriers for meson's adoption in the real world,
  where persuasive individuals do sadly end up expressing distate of its
  "ugly cmake-style syntax".

Due to existing built-in options like --libdir -> -Dlibdir, there *is*
some precedent to supplying a nicer "front-end" options where we can.