opts: Replace dashes by underscores in option names
This normalization was motivated by the discussion in PR #7429.

Due to the added risk of accidentally defining duplicate options, we
also make sure to detect duplicate options and error when they are
present. (Something that is not being done currently).

With the way this normalization was implemented, we now also error on
-D'illegal option name' and get_option('illegal option name'), which
previously both did not error.